chore(useTextarea): import helper functions to declutter hook
